+++ alvr/client_core/src/platform/android/mod.rs
@@ -61,32 +61,6 @@ pub fn try_get_permission(permission: &str) {
     }
 }
 
-pub fn build_string(ty: &str) -> String {
-    let vm = vm();
-    let mut env = vm.attach_current_thread().unwrap();
-
-    let jname = env
-        .get_static_field("android/os/Build", ty, "Ljava/lang/String;")
-        .unwrap()
-        .l()
-        .unwrap();
-    let name_raw = env.get_string((&jname).into()).unwrap();
-
-    name_raw.to_string_lossy().as_ref().to_owned()
-}
-
-pub fn device_name() -> String {
-    build_string("DEVICE")
-}
-
-pub fn model_name() -> String {
-    build_string("MODEL")
-}
-
-pub fn manufacturer_name() -> String {
-    build_string("MANUFACTURER")
-}
-
 fn get_system_service<'a>(env: &mut JNIEnv<'a>, service_name: &str) -> JObject<'a> {
     let service_str = env.new_string(service_name).unwrap();
