+++ alvr/client_core/src/platform/mod.rs
@@ -1,6 +1,2 @@
-#[cfg(not(target_os = "android"))]
-pub fn local_ip() -> std::net::IpAddr {
-    use std::net::{IpAddr, Ipv4Addr};
-
-    local_ip_address::local_ip().unwrap_or(IpAddr::V4(Ipv4Addr::UNSPECIFIED))
-}
+#[cfg(target_os = "android")]
+pub mod android;
