+++ alvr/common/Cargo.toml
@@ -13,6 +13,7 @@ enable-messagebox = ["rfd"]
 anyhow = { version = "1", features = ["backtrace"] }
 backtrace = "0.3"
 glam = { version = "0.29", features = ["serde"] }
+local-ip-address = "0.6"
 log = "0.4"
 once_cell = "1"
 parking_lot = "0.12"
