+++ alvr/common/src/platform/mod.rs
@@ -184,3 +184,10 @@ pub fn platform() -> Platform {
         Platform::Unknown
     }
 }
+
+#[cfg(not(target_os = "android"))]
+pub fn local_ip() -> std::net::IpAddr {
+    use std::net::{IpAddr, Ipv4Addr};
+
+    local_ip_address::local_ip().unwrap_or(IpAddr::V4(Ipv4Addr::UNSPECIFIED))
+}
